Move getQuery to shared utils folder
