+++ assets/js/utils/get-query.js
@@ -1,0 +1,23 @@
+export default function getQuery( attributes ) {
+	const { categories, columns, orderby, rows } = attributes;
+
+	const query = {
+		category: categories.join( ',' ),
+		per_page: rows * columns,
+	};
+
+	if ( 'price_desc' === orderby ) {
+		query.orderby = 'price';
+		query.order = 'desc';
+	} else if ( 'price_asc' === orderby ) {
+		query.orderby = 'price';
+		query.order = 'asc';
+	} else if ( 'title' === orderby ) {
+		query.orderby = 'title';
+		query.order = 'asc';
+	} else {
+		query.orderby = orderby;
+	}
+
+	return query;
+}
